Alterações sigificativas na estrutura do projeto: #3
