--- .gitignore
@@ -1,5 +1,0 @@
-/node_modules
-.env
-
-package-lock.json
-yarn.lock
