--- .prettierrc
@@ -1,4 +1,0 @@
-{
-    "singleQuote": true,
-    "trailingComma": "es5"
-  }
