+++ README.md
@@ -7,14 +7,11 @@
 - Desafio Semanal ( Infos sobre o desafio e o botão de participar )
 - Perfil ( Infos do perfil )
 
-
-
 ### TODO (O que fazer):
 
 - Plataforma com login do Github(Loga com o github) você cria o repositório do projeto lá na plataforma mas o upload dos arquivos pelo git mesmo.
 
 - Usar modelo MERN (MongoDB, Express, React + React Native { Só pra criar os repositórios e Verificar a pontuação } ).
-
 
 ### REQUISITOS FUNCIONAIS
 - O usuário deve poder fazer login pelo Github
