+++ api/.eslintrc.js
@@ -1,0 +1,22 @@
+module.exports = {
+    "env": {
+        "browser": true,
+        "commonjs": true,
+        "es6": true
+    },
+    "extends": "airbnb-base",
+    "globals": {
+        "Atomics": "readonly",
+        "SharedArrayBuffer": "readonly"
+    },
+    "parserOptions": {
+        "ecmaVersion": 2018,
+        "sourceType": "module"
+    },
+    "rules": {
+      "no-console": "off",
+      "max-len": "off",
+      "prefer-destructuring": "off",
+      "consistent-return": "off"
+    }
+};
