+++ api/.gitignore
@@ -1,0 +1,16 @@
+build/
+# Commonly ignored Node.js files
+node_modules/
+npm-debug.log
+.npm/
+
+# OSX system files, the bane of our existence
+.DS_Store
+.AppleDouble
+.LSOverride
+
+# DB Connection String
+config/
+
+package-lock.json
+yarn.lock
