+++ api/src/Database/index.js
@@ -1,0 +1,6 @@
+import { stringConnection } from '../configs';
+import mongoose from 'mongoose';
+
+mongoose.connect(stringConnection, {
+  useNewUrlParser: true
+});
