--- nodemon.json
@@ -1,5 +1,0 @@
-{
-	"execMap": {
-		"js": "sucrase-node"
-	}
-}
