--- package.json
@@ -1,32 +1,0 @@
-{
-  "name": "plataforma-desafios-js-api",
-  "version": "1.0.0",
-  "main": "src/index.js",
-  "scripts": {
-    "dev": "nodemon src/index.js",
-    "queue": "nodemon src/queue.js"
-  },
-  "repository": "https://github.com/b3coded/Plataforma-Desafios-JS-API.git",
-  "author": "b3coded",
-  "license": "MIT",
-  "devDependencies": {
-    "eslint": "^5.16.0",
-    "eslint-config-airbnb-base": "^13.1.0",
-    "eslint-config-prettier": "^6.0.0",
-    "eslint-plugin-import": "^2.18.0",
-    "eslint-plugin-prettier": "^3.1.0",
-    "nodemon": "^1.19.1",
-    "prettier": "^1.18.2",
-    "sucrase": "^3.10.1"
-  },
-  "dependencies": {
-    "bcrypt": "^3.0.6",
-    "cors": "^2.8.5",
-    "dotenv": "^8.0.0",
-    "express": "^4.17.1",
-    "jsonwebtoken": "^8.5.1",
-    "mongoose": "^5.6.0",
-    "socket.io": "^2.2.0",
-    "yup": "^0.27.0"
-  }
-}
