--- src/Database/index.js
@@ -1,6 +1,0 @@
-import { stringConnection } from '../configs';
-import mongoose from 'mongoose';
-
-mongoose.connect(stringConnection, {
-  useNewUrlParser: true
-});
